third_party/blink/web_tests/virtual/document-pip/document-pip-clears-session-on-close.html

<!DOCTYPE html>
<title>Test that navigator.documentPictureInPicture.session
  is cleared when the PiP window in closed.</title>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../../resources/user-gesture-utils.js"></script>
<body>
  <div id="div"></div>
<script>
const div = document.getElementById('div');

async_test((t) => {
  focusWithUserGesture(window);
  navigator.documentPictureInPicture.requestWindow().then( () => {
    var session = navigator.documentPictureInPicture.session;
    assert_true(!!session, "session should exist with PiP is opened");
    session.window.addEventListener('unload', () => {
      // `session` is still set until event processing is complete.
      scheduler.postTask(t.step_func_done( () => {
        assert_true(!navigator.documentPictureInPicture.session,
                    "session should be cleared when PiP is closed");
      }))
    });
    session.window.close();
  });
});
</script>
</body>
